updating healthchek type path
