+++ src/shared/types/api.type.ts
@@ -14,34 +14,6 @@ export type ApiError = {
   resultMessage: string;
 };
 
-//healthcheck types:
-export type AppHealthDetails = {
-  timestamp: string;
-  uptime: number;
-  memoryUsage: {
-    total: number;
-    free: number;
-  };
-  startTime: number;
-};
-
-export type DbHealthDetails = {
-  timestamp: string;
-  connectionTime?: number;
-  databaseStatus?: string;
-  error?: string;
-};
-
-export type HealthCheckProps<T> = {
-  status: string;
-  message: string;
-  details?: T;
-};
-
-export type HealthCheckData<T> = {
-  healthCheck: HealthCheckProps<T>;
-};
-
 //pagination and task related ops
 
 export type PaginationMetadata = {
